+++ library/Freckle/App/Bugsnag.hs
@@ -23,6 +23,7 @@ module Freckle.App.Bugsnag
 import Freckle.App.Prelude
 
 import Control.Lens (Lens', view)
+import Control.Monad.Catch (MonadMask)
 import Control.Monad.Reader (runReaderT)
 import Data.Bugsnag
 import Data.Bugsnag.Settings
@@ -32,7 +33,7 @@ import Database.PostgreSQL.Simple (SqlError (..))
 import Database.PostgreSQL.Simple.Errors
 import Freckle.App.Async (async)
 import qualified Freckle.App.Env as Env
-import qualified Freckle.App.Exception as Exception
+import qualified Freckle.App.Exception.MonadUnliftIO as Exception
 import Network.Bugsnag hiding (notifyBugsnag, notifyBugsnagWith)
 import qualified Network.Bugsnag as Bugsnag
 import Network.HTTP.Client (HttpException (..), host, method)
